+++ internal/db/models.go
@@ -59,7 +59,7 @@ type ClossCustomer struct {
 	Prcdpagdia      float64
 	Promdiasp       float64
 	Riesgocrd       float64
-	Cantdocs        float64
+	Cantdocs        int64
 	Totmtodocs      float64
 	Prommtodoc      float64
 	Diasultvta      float64
